Fix parent discarder regexp
